Move to using builds/ & downloads/
This switches from using the top-level base directory for downloads
& builds to putting them into a more logical place: downloads & builds.
This hasn't been fully tested, but seems to work so far.

This also includes an update to the location of where nuttx is
downloaded from. However, there appears to be a bug in the github tar
downloads for nuttx that packages them as: nuttx-nuttx-$version.
A workaround has been added for this.